OS/2 personality enhancements - Move most RPC functions from os2app to os2srv. Now, for example, shared memory dataspaces are created by os2srv and one os2ap copy asks os2srv to map it to another os2app copy, instead of calling RPC for other os2app copy to share it. - Use PID's instead of native thread id in some os2srv RPC functions. - Further portability enhancements. - More system-dependent functions implemented for Genode platform. - trampoline() assembly code moved to separate tramp() function, which has implementations for x86_32 and x86)64. - Enhance system-dependent functions is compat lib for Genode platform. - Add DosFreeModule implementation.

+++ OS2/Server/include/os3/memmgr.h
@@ -27,7 +27,7 @@
 
 struct t_mem_area {
 	void * start;
-	unsigned int size;
+	unsigned long size;
 	void * next;
 };
 
@@ -35,9 +35,9 @@ struct t_mem_area {
 
 void init_memmgr(struct t_mem_area *root_area);
 
-void * seek_free_mem(struct t_mem_area *root_area, unsigned int i_size);
-int   is_mem_used(struct t_mem_area *root_area, void * p_start, unsigned int i_size);
-int   alloc_mem_area(struct t_mem_area *root_area, void * p_start, unsigned int i_size);
+void * seek_free_mem(struct t_mem_area *root_area, unsigned long i_size);
+int   is_mem_used(struct t_mem_area *root_area, void * p_start, unsigned long i_size);
+int   alloc_mem_area(struct t_mem_area *root_area, void * p_start, unsigned long i_size);
 int   dealloc_mem_area(struct t_mem_area *root_area, void * p_start);
 void  print_used_mem(struct t_mem_area *root_area);
